+++ docs/changelog.rst
@@ -29,6 +29,7 @@ Fixed
 
 * Fix internal server error when replying to content that contained only characters outside the western Latin character sets.
 * Visual fixes for content rendering in content delete page.
+* Make direct profile handle search survive extra spaces before or after the searched handle.
 
 0.2.1 (2017-07-30)
 ------------------